% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedAlgo.R
\name{bedAlgo}
\alias{bedAlgo}
\title{bedAlgo provides the continuous probability of an event at every prediction timestep.
 The memory about the recent past event is automatically included to predict the future.
 An event is initiated when the BED probability exceeds the eventThreshhold.}
\usage{
bedAlgo(model)
}
\arguments{
\item{model}{model to which the postprocessor shall be added}
}
\value{
model model with added postprocessing results
}
\description{
bedAlgo provides the continuous probability of an event at every prediction timestep.
 The memory about the recent past event is automatically included to predict the future.
 An event is initiated when the BED probability exceeds the eventThreshhold.
}
\author{
Sowmya
}
\keyword{internal}
