% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildEDModel}
\alias{buildEDModel}
\title{build Event Detection Model}
\usage{
buildEDModel(x, dataPrepators = "ImputeTSInterpolation",
  dataPreparationControl = list(), buildModelAlgo = "ForecastETS",
  buildModelControl = list())
}
\arguments{
\item{x}{data}

\item{dataPrepators}{string or vector of strings, which preparators to use}

\item{dataPreparationControl}{control list for data preparators}

\item{buildModelAlgo}{string name of modelling algo}

\item{buildModelControl}{control list for modelling algo}
}
\value{
model fittedModel
}
\description{
build Event Detection Model
}
