% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorsAndWarnings.R
\name{edError}
\alias{edError}
\title{Generate a Package EventDetectR specific Error}
\usage{
edError(strError)
}
\arguments{
\item{strError}{Message to print}
}
\description{
Generate a Package EventDetectR specific Error
}
\keyword{internal}
