% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorsAndWarnings.R
\name{edWarning}
\alias{edWarning}
\title{Generate a Package EventDetectR specific Warning}
\usage{
edWarning(strError)
}
\arguments{
\item{strError}{Message to print}
}
\description{
Generate a Package EventDetectR specific Warning
}
\keyword{internal}
