% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventClassification.R
\name{eventClassification}
\alias{eventClassification}
\title{Predict Univariate Models Forecast Package}
\usage{
eventClassification(object, newData, ...)
}
\arguments{
\item{object}{fitted model that shall be predicted}

\item{...}{additional parameters}
}
\value{
eventDetector fittedModel with
}
\description{
Predict Univariate Models Forecast Package
}
\keyword{internal}
