% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventDetectR-package.R
\docType{package}
\name{EventDetectR-package}
\alias{EventDetectR-package}
\title{EventDetectR-package description}
\description{
The EventDetectR package is a modular event detection system
}
\details{
The EventDetectR package enables detection of events/ anomalies in multivariate time-series data. It combines multiple well-known R packages like 'forecast, 'neuralnet' to deliver an easily configurable tool for event detection.
}
