% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDocumentation.R
\docType{data}
\name{geccoIC2018Test}
\alias{geccoIC2018Test}
\title{geccoIC2018Test}
\description{
2018s Test set of the gecco industrial challenge - http://www.spotseven.de/gecco/gecco-challenge/
}
\keyword{data}
