% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultControlLists.R
\name{getDefaultForecastModelControl}
\alias{getDefaultForecastModelControl}
\title{Get Default Control List for ED - Forecast Model Building Algorithms}
\usage{
getDefaultForecastModelControl()
}
\value{
defaultControlList List with named arguments, 1 for each default parameter
}
\description{
Get Default Control List for ED - Forecast Model Building Algorithms
}
\keyword{internal}
