% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultControlLists.R
\name{getDefaultModelControl}
\alias{getDefaultModelControl}
\title{Get Default Control List for ED - Model Building Algorithms}
\usage{
getDefaultModelControl()
}
\value{
defaultControlList List with named arguments, 1 for each default parameter
}
\description{
Get Default Control List for ED - Model Building Algorithms
}
\keyword{internal}
