% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultControlLists.R
\name{getDefaultNeuralNetModelControl}
\alias{getDefaultNeuralNetModelControl}
\title{Get Default Control List for ED - NN Model Building Algorithms}
\usage{
getDefaultNeuralNetModelControl()
}
\value{
defaultControlList List with named arguments, 1 for each default parameter
}
\description{
Get Default Control List for ED - NN Model Building Algorithms
}
\keyword{internal}
