% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultControlLists.R
\name{getDefaultPostControl}
\alias{getDefaultPostControl}
\title{Get Default Control List for ED - PostProcessors}
\usage{
getDefaultPostControl()
}
\value{
defaultControlList List with named arguments, 1 for each default parameter
}
\description{
Get Default Control List for ED - PostProcessors
}
\keyword{internal}
