% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDefaultControlLists.R
\name{getDefaultPreparationControl}
\alias{getDefaultPreparationControl}
\title{Get Default Control List for ED - Data Preparators}
\usage{
getDefaultPreparationControl()
}
\value{
defaultControlList List with named arguments, 1 for each default parameter
}
\description{
Get Default Control List for ED - Data Preparators
}
\keyword{internal}
