% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedMethods.R
\name{getSupportedModels}
\alias{getSupportedModels}
\title{getSupportedModels}
\usage{
getSupportedModels()
}
\value{
allSupportedModels a list of strings with each supported method name. The strings can be copied and used in calls to 'eventDetect' or 'buildEDModel'
}
\description{
Get a list of all data modelling methods that are currently supported in package 'eventDetectR'.
}
\examples{
models <- getSupportedModels()
}
