% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedMethods.R
\name{getSupportedPostProcessors}
\alias{getSupportedPostProcessors}
\title{getSupportedPostProcessors}
\usage{
getSupportedPostProcessors()
}
\value{
allSupportedPostProcessors a list of strings with each supported method name. The strings can be copied and used in calls to 'eventDetect' or 'buildEDModel'
}
\description{
Get a list of all data postprocessing methods that are currently supported in package 'eventDetectR'.
}
\examples{
preps <- getSupportedPostProcessors()
}
