% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSupportedMethods.R
\name{getSupportedPreparations}
\alias{getSupportedPreparations}
\title{getSupportedPreparations}
\usage{
getSupportedPreparations()
}
\value{
allSupportedPreparations a list of strings with each supported method name.
The strings can be copied and used in calls to 'eventDetect' or 'buildEDModel'
}
\description{
Get a list of all data preparation methods that are currently supported in package 'eventDetectR'.
}
\examples{
preps <- getSupportedPreparations()
}
