% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_NeuralNetwork.R
\name{model_NeuralNetwork}
\alias{model_NeuralNetwork}
\title{Fitting Neuralnet Models}
\usage{
model_NeuralNetwork(x, control)
}
\arguments{
\item{x}{data}

\item{control}{control list with settings}
}
\value{
fitted multivariate neural network model
}
\description{
Fitting Neuralnet Models
}
\details{
The Neuralnetworks are used to model the multivariate parameters through non-linear, weighted, parametrized functions. A Neuralnetwork model is formulated for each of the parameter using its own lagged values and rest of all parameters involved.
}
\author{
Sowmya
}
\keyword{internal}
