% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Forecast.R
\name{model_UnivariateForecast}
\alias{model_UnivariateForecast}
\title{Fitting Forecast Models}
\usage{
model_UnivariateForecast(x, strName, control)
}
\arguments{
\item{x}{data}

\item{strName}{name of the Forecast model that shall be fitted}

\item{control}{control list with settings}
}
\value{
fitted Forecast model
}
\description{
Fitting Forecast Models
}
\keyword{internal}
