% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvents.R
\name{plot.edObject}
\alias{plot.edObject}
\title{Plot an Event Detection Object}
\usage{
\method{plot}{edObject}(x, varsToPlot = names(edObject$classification),
  ...)
}
\arguments{
\item{x}{edObject}

\item{varsToPlot}{vars}

\item{...}{Additional parameters}
}
\value{
A Plot
}
\description{
Plot an Event Detection Object
}
