% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_NeuralNetwork.R
\name{predict.NeuralNetwork}
\alias{predict.NeuralNetwork}
\title{Predict MultivariateNeuralNetwork Models neuralnet Package}
\usage{
\method{predict}{NeuralNetwork}(object, newData = NULL, ...)
}
\arguments{
\item{object}{fitted model that shall be predicted}

\item{newData}{data.frame with newData that is compared to the models prediction to judge if events occured or not}

\item{...}{additional parameters}
}
\value{
predicted value
}
\description{
Predict MultivariateNeuralNetwork Models neuralnet Package
}
\keyword{internal}
