% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Forecast.R
\name{predict.UnivariateForecast}
\alias{predict.UnivariateForecast}
\title{Predict Univariate Models Forecast Package}
\usage{
\method{predict}{UnivariateForecast}(object, ...)
}
\arguments{
\item{object}{fitted model that shall be predicted}

\item{...}{additional parameters}
}
\value{
predicted value
}
\description{
Predict Univariate Models Forecast Package
}
