% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparator_imputeTS.R
\name{preparator_imputeTS}
\alias{preparator_imputeTS}
\title{Data Preparation with Impute TS}
\usage{
preparator_imputeTS(x, prepStr, control)
}
\arguments{
\item{x}{data to process}

\item{prepStr}{string of preparator name to use}

\item{control}{control parameters, extra settings}
}
\value{
x prepared data
}
\description{
Data Preparation with Impute TS
}
\keyword{internal}
