% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvents.R
\name{print.edObject}
\alias{print.edObject}
\title{Print an Event Detection Object}
\usage{
\method{print}{edObject}(edObject, nLast = 10)
}
\arguments{
\item{edObject}{edObject, the event detection object that shall be printed}

\item{nLast}{integer, last n rsults that shall be printed}
}
\description{
Prints the last classification results for an event detection object.
}
