% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateEvents.R
\name{simulateEvents}
\alias{simulateEvents}
\title{Generates simulated events}
\usage{
simulateEvents(Data, Params, Event_type, Event_strength = NULL,
  Start_index, Event_duration = NULL, Percentage = NULL)
}
\arguments{
\item{Data}{Data frame or matrix containing the data which will undergo the transformations.}

\item{Params}{Vector or vector of strings indicating the column names (in case Data is a data frame) or the column
numbers (in case Data is a matrix) of the parameters in which an event will be simulated}

\item{Event_type}{String vector indicating which type of transformation the parameters will undergo.
Current valid options include sinusoidal, square, ramp, binomial and slowsinusoidal. If Params contains more that one
elements and Event_type only contains one element the same transformation will be applied to all given Params}

\item{Event_strength}{(Optional) Numeric Vector indicating the amplitude. Only valid for sinusoidal and square
transformations}

\item{Start_index}{Index indicated where the event should start}

\item{Event_duration}{Numeric, indicates the number of steps the transformation should last. Default is 100}

\item{Percentage}{Numeric value from 0 to 1. Alternative input indicating the percentage of data that
should be affected by the transformation. Either Event_duration or Percentage should be especified.}
}
\value{
Matrix or data frame containing the transformed columns
}
\description{
Simulates Events on columns of a data frame or a matrix by appliying different transformations.
The events can be of type sinusoidal,
square, binomial or ramp.
}
\examples{

#Generate event of type sinusoidal and binomial on two columns of the stationBData data set
simupar<-c("B_PH_VAL","B_TEMP_VAL")
SimulatedEvents<-simulateEvents(stationBData,simupar,Event_type = c("sinusoidal","ramp"), Start_index = 2500)

}
