% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_station_b.R
\docType{data}
\name{test_station_b}
\alias{test_station_b}
\title{Example data}
\format{Time Series (\code{\link{ts}}) with 144 rows including 13 NAs.}
\source{
\cite{hallo}
}
\usage{
test_station_b
}
\description{
Description example data
}
\details{
The dataset originates from ...
}
\keyword{datasets}
