% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedAlgo.R
\name{bedAlgo}
\alias{bedAlgo}
\title{Dummy BED algo, has to be changed}
\usage{
bedAlgo(model)
}
\arguments{
\item{model}{model to which the postprocessor shall be added}
}
\value{
model model with added postprocessing step
}
\description{
Dummy BED algo, has to be changed
}
\keyword{internal}
