% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEDModel.R
\name{buildEDModel}
\alias{buildEDModel}
\title{build Event Detection Model}
\usage{
buildEDModel(x, dataPrepators = "ImputeTSInterpolation",
  dataPreparationControl = list(), buildModelAlgo = "ForecastETS",
  buildModelControl = list(), postProcessors = "bedAlgo",
  postProcessorControl = list())
}
\arguments{
\item{x}{data}

\item{dataPrepators}{string or vector of strings, which preparators to use, !no list!}

\item{dataPreparationControl}{control list for data preparators}

\item{buildModelAlgo}{string name of modelling algo}

\item{buildModelControl}{control list for modelling algo}

\item{postProcessors}{string name of one or more postprocessors, !no list!}

\item{postProcessorControl}{control list for postprocessors}
}
\value{
model fittedModel
}
\description{
build Event Detection Model
}
