% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventDetectR-package.R
\docType{package}
\name{EventDetectR-package}
\alias{EventDetectR-package}
\title{EventDetectR-package description}
\description{
The EventDetectR package is a modular event detection system
}
\details{
The EventDetectR package is very good, you can use it like this...
}
\references{
...
}
