% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_NeuralNetwork.R
\name{model_NeuralNetwork}
\alias{model_NeuralNetwork}
\title{Fitting Neuralnet Models}
\usage{
model_NeuralNetwork(x, control)
}
\arguments{
\item{x}{data}

\item{control}{control list with settings}
}
\value{
fitted multivariate neural network model
}
\description{
Fitting Neuralnet Models
}
\details{
Multivariate Neuralnetwork model is formulated for each of the parameter using its own lagged values and rest of all parameters involved. For instance, considering 3 parameters 'p1', 'p2', 'p3', the value of 'p1' at time step 'i' is calclated as \cr
\code{p1[i] = p2[i] + p3[i] + p1[i-1]} \cr
}
\keyword{internal}
