% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Forecast.R
\name{predict.UnivariateForecast}
\alias{predict.UnivariateForecast}
\title{Predict Univariate Models Forecast Package}
\usage{
\method{predict}{UnivariateForecast}(object, newData = NULL, ...)
}
\arguments{
\item{object}{fitted model that shall be predicted}

\item{newData}{data.frame with newData that is compared to the models forecast to judge if events occured or not}

\item{...}{additional parameters}
}
\value{
predicted value
}
\description{
Predict Univariate Models Forecast Package
}
\keyword{internal}
