% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectEvents.R
\name{print.edObject}
\alias{print.edObject}
\title{Print an Event Detection Object}
\usage{
\method{print}{edObject}(x, ...)
}
\arguments{
\item{x}{edObject, the event detection object that shall be printed}

\item{...}{any additional parameters}
}
\description{
Prints the last classification results for an event detection object.
If 'nLast' (integer) is given, it specifies the amount of rows to be printed.
}
